Make the theme available to the rest of the app
