+++ client/src/GlobalProvider.js
@@ -1,27 +1,11 @@
 import React, { Component } from "react";
 
-import { MuiThemeProvider, createMuiTheme } from '@material-ui/core/styles';
+import { MuiThemeProvider } from '@material-ui/core/styles';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
 
+import theme from './MuiTheme';
 import App from './App';
-
-const theme = createMuiTheme({
-	palette: {
-		primary: {
-			light: '#5472d3',
-			main: '#0d47a1',
-			dark: '#002171',
-			contrastText: '#fff',
-		},
-		secondary: {
-			light: '#60ad5e',
-			main: '#2e7d32',
-			dark: '#005005',
-			contrastText: '#ffffff',
-		},
-	},
-});
 
 class GlobalProvider extends Component {
 	render() {
